Add tests for cinder nimblestorage and solidfire
Turn netapp tests into parent class CinderBackendTest.
Solidfire, nimblestorage and netapp inheret tests from parent.
